Divide CreateBrep to some cs files to improve its maintainability
